* Fix hash type to be 64bits on most platforms (hopefully) * Add tests for base urls

+++ src/list.h
@@ -4,7 +4,7 @@
 struct node {
   bstring name;
   bstring content;
-  long level;
+  long long level;
   struct node *next;
 };
